refactor(roles): move to config folder
